catalog onboarding (#10)
