+++ extensions/landing-zone/catalogValidationValues.json.template
@@ -1,4 +1,3 @@
 {
-  "ibmcloud_api_key": $VALIDATION_APIKEY,
-  "cluster_id": $CLUSTER_ID
+  "ibmcloud_api_key": $VALIDATION_APIKEY
 }
