+++ tests/resources/outputs.tf
@@ -7,9 +7,21 @@ output "prefix" {
   description = "prefix"
 }
 
-output "cluster_id" {
+output "region" {
+  value       = var.region
+  description = "Region where SLZ ROKS Cluster is deployed."
+}
+
+output "management_cluster_id" {
   value = lookup(
-    [for cluster in module.landing_zone.cluster_data : cluster][0], "cluster_id", ""
+    [for cluster in module.landing_zone.cluster_data : cluster if strcontains(cluster.resource_group_name, "management")][0], "id", ""
   )
-  description = "cluster ID"
+  description = "management cluster ID"
+}
+
+output "workload_cluster_id" {
+  value = lookup(
+    [for cluster in module.landing_zone.cluster_data : cluster if strcontains(cluster.resource_group_name, "workload")][0], "id", ""
+  )
+  description = "workload cluster ID"
 }
