+++ tests/scripts/pre-validation-slz-roks.sh
@@ -21,33 +21,22 @@ TF_VARS_FILE="terraform.tfvars"
   # $VALIDATION_APIKEY is available in the catalog runtime
   {
     echo "ibmcloud_api_key=\"${VALIDATION_APIKEY}\""
-    echo "prefix=\"was-slz-$(openssl rand -hex 2)\""
     echo "region=\"${REGION}\""
   } >> ${TF_VARS_FILE}
   terraform apply -input=false -auto-approve -var-file=${TF_VARS_FILE} || exit 1
 
-  prefix_var_name="prefix"
-  slz_prefix_var_name="landing_zone_prefix"
-  prefix_var_value=$(terraform output -state=terraform.tfstate -raw prefix)
-  rg_var_name="resource_group"
-  rg_var_value="${prefix_var_value}-management-rg"
   region_var_name="region"
   cluster_id_var_name="cluster_id"
-  cluster_id_value=$(terraform output -state=terraform.tfstate -raw cluster_id)
+  cluster_id_value=$(terraform output -state=terraform.tfstate -raw management_cluster_id)
 
-  echo "Appending '${prefix_var_name}', '${cluster_id_var_name}', '${rg_var_name}' and '${region_var_name}' input variable values to ${JSON_FILE}.."
+  echo "Appending '${cluster_id_var_name}' and '${region_var_name}' input variable values to ${JSON_FILE}.."
 
   cd "${cwd}"
-  jq -r --arg prefix_var_name "${prefix_var_name}" \
-        --arg slz_prefix_var_name "${slz_prefix_var_name}" \
-        --arg prefix_var_value "${prefix_var_value}" \
-        --arg rg_var_name "${rg_var_name}" \
-        --arg rg_var_value "${rg_var_value}" \
-        --arg region_var_name "${region_var_name}" \
+  jq -r --arg region_var_name "${region_var_name}" \
         --arg region_var_value "${REGION}" \
         --arg cluster_id_var_name "${cluster_id_var_name}" \
         --arg cluster_id_value "${cluster_id_value}" \
-        '. + {($prefix_var_name): $prefix_var_value, ($slz_prefix_var_name): $prefix_var_value, ($rg_var_name): $rg_var_value, ($region_var_name): $region_var_value}, ($cluster_id_var_name): $cluster_id_value' "${JSON_FILE}" > tmpfile && mv tmpfile "${JSON_FILE}" || exit 1
+        '. + {($region_var_name): $region_var_value}, ($cluster_id_var_name): $cluster_id_value' "${JSON_FILE}" > tmpfile && mv tmpfile "${JSON_FILE}" || exit 1
 
   echo "Pre-validation complete successfully"
 )
